rename: entity 패키지 명 수정 (-> model)
