+++ src/main/java/com/devleoh/glomo/asset/model/Asset.java
@@ -1,4 +1,4 @@
-package com.devleoh.glomo.asset.entity;
+package com.devleoh.glomo.asset.model;
 
 import com.devleoh.glomo.base.BaseEntity;
 import jakarta.persistence.*;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 
 /**
  * packageName    : com.devleoh.glomo.product.entity
- * fileName       : TradingMarket
+ * fileName       : Product
  * author         : nimoh
  * date           : 2024/08/01
  * description    :
@@ -19,13 +19,23 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor
-public class TradingMarket extends BaseEntity {
+public class Asset extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="TRADING_MARKET_ID")
-    private Long tradingMarketId;
+    @Column(name = "ASSET_ID")
+    private Long assetId;
 
     @Column(nullable = false)
-    private String name;
+    private String code;                        // 상품 코드
+
+    private int faceValue;                      // 액면가
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="ASSET_CATEGORY_ID")
+    private AssetCategory assetCategory;    // 상품 분류
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="TRADING_MARKET_ID")
+    private TradingMarket tradingMarket;        // 시장
 }
